// исходный код/mf-core-main/mf-core-main/src/stores/Corporate/Corporate.interface.ts
export interface ICorporateStore {
  selfEmployee: EmployeeModel | undefined;
  organizations: IOrganization[];
  departments: DepartmentModel[];
  positions: IPosition[];
  organizationsMapped: Dictionary<OrganizationNormalizedModel>;
  departmentsMapped: Dictionary<TDepartment>;
  positionsMapped: Dictionary<IPosition>;
  savedDepartment?: DepartmentDetailedModel;
  initStore(): void;
  clearSavedDepartment(): void;
  editSavedDepartment(model: DepartmentDetailedModel): void;
  refreshDepartments(): void;

  getDepartment(depId: string): TDepartment;
  editDepartment(model: DepartmentDetailedModel): void;
  deleteDepartment(depId: string): void;
}
// ...
export interface IOrganization {
  id: string;
  officialName: string;
  address: string;
  positions: IShortPosition[];
  departments: IShortDepartment[];
}

export interface IShortDepartment {
  id: string;
  departmentName: string;
}
// ...
export const IODepartment: t.Type<IDepartment> = t.recursion('IODepartment', () => t.intersection([
  t.type({
    organizationId: t.string,
    code: t.string,
    departmentName: t.string,
    fullStructurePath: t.string,
    employees: t.array(IODepartmentEmployee),
    children: t.array(IODepartment),
    status: ioTypeFromEnum<EmployeeStatus>('EmployeeStatus', EmployeeStatus),
  }),
  t.partial({
    id: t.string,
    humanReadableId: t.string,
    departmentHeadId: t.string,
    parentId: t.string,
    location: t.string,
  }),
])
);

export type TDepartment = t.TypeOf<typeof IODepartment>;
// ...
export interface IPosition {
  id: string;
  organizationId: string;
  positionName: string;
  selfApproved: boolean;
  availableClasses: TaxiClassEnum[];
}

export interface IShortPosition {
  id: string;
  organizationId: string;
  positionName: string;
}